tools: dtnstream improved with flow control
This patch adds automatic flow control to the streaming source
using the StatusReport features. On each received status report
the existing buffer is released.

In addition, the buffer handling has been improved so that the
"-c" parameter defines the minimum bundle size (buffer size)
and "-C" the maximum bundle size. The maximum buffer size if
even valid, if flow control is activated.