ibrdtn: Fix segfault in BundleList caused by bundles with an AgeBlock
The segfault occur if bundles with an unprecise expiration timestamp are
stored in a BundleList. A unprecise expiration time is the result of bundles
with an AgeBlock or if the local clock is marked as non-perfect. In such a
case, bundles might not removed correctly from the internal structure of
the BundleList which leads to a segfault during the expiration process.

This patch solves the issue by using the exact same object for the insertion
and removal.