daemon: Fix bundle transfer release bug
If neighbor was unavailable and a transfer was completed in the meantime, the transfer
was not released in the neighbor entry. This patch modifies the get() method of the
neighbor database to return cached neighbor entries even if the neighbor is not available.
For those callers who need the old behavior, an additional parameter has been introduced
which must be set to true in those cases.