Add fragment length to unique bundle ID
As described in the RFC 4838 the bundle fragment length is part of the unique bundle
identifier. This patch adds the fragment length to the BundleID object and changes the
payload length field in the status reports and custody signals to that value. The
serialization of the BundleID is also modified to include that value.