Whisper: Fix crash on disconnect from 0.12 router
When trunk Whisper is run with 0.12 IBR-DTN router, it crashes when the
contact list or message list windows are closed. This is because
binding the security connection fails. The bind exception is caught,
but then attempting to unbind it causes an unrecoverable error. Track
the connection bind status so Whisper does not attempt to unbind an
unbound connection.