ibrdtn: Fix segfault if BLOB creation fails
The signature of BundleBuilder::insert() was restricted to throw only InvalidDataException
instances. The method actually instantiate Block instances and therefore also BLOB
instances. In case the creation of the BLOB fails due to a full disk, the insert() method
throws a IOException. For that reason, the throw() restriction is removed completely to
allow also IOException instances to be thrown.