ibrdtn: Make non-reference clocks independent from outer clock adjustments
Before this patch it was possible to disturb the time-synchronization process by
setting the hosts clock individually. Now, a reference is recorded at the start and
adjustments are applied to this value. The current time-stamp is then calculated using
the BOOT_TIME / MONOTONIC_TIME of the host and the previously stored reference.
That way it is not possible change internal clock by setting the host time.