daemon: Make beacon interval configurable
The option discovery_interval replaces discovery_timeout and specifies how often
discovery beacons are sent. By default every 5 seconds instead of each second. In
suspend mode, this value is increased by a factor of ten.

For incoming beacons without a timeout value, the value of the option is also considered
to define a timeout value.