All the modifications described below aim at supporting emulation with IBRDTN nodes. In emulation mode, several instances of IBRDTN can run concurrently on the same host, and all traffic is sent to a software hub, which forwards messages to their actual destination(s). This sofware hub can itself be controlled by a mobility simulator, in which case each message is only forwarded by the hub if the sender is considered as a neighbour of the destination at the time the message is received by the hub.
LEPTON (http://www-casa.irisa.fr/lepton) is a lightweight emulation
platform for opportunistic networking that works along that line.

In order for IBRDTN nodes to support emulation mode, the original
IBRDTN code must be slightly modified. The modifications described
below mostly pertain to UDP-based beaconing and to the TCP convergence
layer.  None of these modifications compromise the normal behaviour of
IBRDTN, though.

------------------------------------------------------------

1) Instead of sending beacons to a multicast group, beacons must be
sent in unicast mode to a static address, which is that of the
software hub. Parameters 'discovery_address' and 'discovery_port' in
the configuration file are sufficient to indicate how to reach the
hub, so no modification of the code is required. In emulation mode,
'discovery_address' is therefore used to specify a unicast
address. IPNDAgent will then complain at startup that it has failed
to join that address. Since this is just a warning in the log files,
this is no big deal... Ideally the join() in IPNDAgent.cpp (line 110)
should be called only if the address is a multicast address.

------------------------------------------------------------

2) Since the hub must be able to forward beacon messages to each
IBRDTN node individually, an IBRDTN node must be able to listen to a
UDP port that may be different from the port number the hub listens
to. An additional parameter 'discovery_local_port' can therefore be
defined in the configuration file, and a new function localPort() is
defined in Configuration.h and implemented in Configuration.cpp to
return the value of that parameter. In NativeDaemon.cpp function
localPort() is called to initialize variable disco_port, before
creating an IPNDAgent.  If parameter 'discovery_local_port' is not
defined in the configuration file, then function localPort() returns
the same value as function port(). This ensures compatibility with the
standard IBRDTN behavior, when emulation mode is not required.

------------------------------------------------------------

3) When several IBRDTN nodes must run concurrently on the same host,
each IPNDAgent should listen to a distinct UDP port. Since during an
experiment dozens or hundreds of nodes can be created and terminated
dynamically (especially when simulating a network with churn), the
port numbers used by these nodes cannot be determined in advance. In
that case it is necessary to use dynamic port allocation.

In IPNDAgent.cpp, function componentUp() has been modified so that if
parameter port (actually disco_port, thus discovery_local_port) has
been set to 0, then dynamic port allocation is used: a dummy socket
named dyn_sock is created and enabled temporarily, with port number
0. One this socket is up, its actual port number used to create all
the other sockets required by IPNDAgent. Once these sockets have been
created, the dummy socket can be closed and deleted. This approach is
a bit tricky, but required because IPNDAgent creates one socket for
every public address of the target interface, and binds each socket to
one address.

A new function get_port() has been defined in socket.h and implemented
in socket.cpp in order to return the local port of a socket, once this
socket is up. This function is used by IPNDAgent to "discover" which
port number has been assigned by the system, once the socket is open.

------------------------------------------------------------

4) When several IBRDTN nodes must run concurrently on the same host,
with the TCP convergence layer enabled, they cannot all listen to the
same TCP port number. So dynamic port allocation is required for TCP
sockets as well.

File TCPConvergenceLayer.cpp has been modified to support dynamic port
assignment, using the same method as in IPNDAgent. When a tcpserver
must be created to listen to port 0 (as specified in the configuration
file), then a dummy socket is first created with dynamic port
assignment. Then the port number assigned to this socket is used as a
reference to create all other sockets.

In socket::tcpserversocket::up(), the socket is bound to an address
only if the port number in that address is != 0. Otherwise the port
number is assigned by the OS, and this port number is "discovered"
(using function get_port()) after we start listening to the
socket. This number is recorded in _address, which therefore cannot be
a 'const' anymore.

------------------------------------------------------------

5) When experiments involving hundreds of IBRDTN nodes are conducted
(especially in a scenario with churn), it is difficult to ensure the
repeatability of these experiments. New parameters have been
introduced in the daemon's configuration in order to provide better
control on the behavior of an IBRDTN node.

In subclass Configuration::Daemon, new parameters start_time,
stop_time, and seed are meant to determine respectively the time when
an IBRDTN daemon should start, the time when it should stop, and the
seed value to be used by any random generator. New options admitted in
the command line allow the user to set these parameters. There is no
side-effect if these parameters are missing.

In Main.cpp, a subclass TerminationThread has been defined that allows
the daemon to terminate spontaneously after a certain execution delay.

In function Main::main(), code has been introduced that sets the
random generator's seed (if option '-s appears in the command line),
that defers the actual start of the daemon (if option '-b' appears in
the command line), and that starts TerminationThread (if option '-e'
appear in the command line).